src/requirements.txt
Flask
Flask-Cors
flask-restx
google-auth
mysql-connector-python
requests
